Added bookmarks+
Also configured it to play nice with workgroups and projectile. Each
project/workgroup now has a specific bookmark file that is automatically
switched to when the project is opened.

The same behaviour was adapted for the way `yas-snippet-dirs` is set
up. It is now set automatically whenever the project/workgroup is
switched (No need to have it in `.dir-locals.el` anymore).